requirements.txt
# các thư viện được sử dụng 
schedule
python-dotenv>=1.0